requirements.txt
easy-pil
aiohttp==3.13.1
aiosqlite==0.21.0
asyncpraw==7.8.1
azure-cognitiveservices-speech==1.46.0
azure-core==1.36.0
discord.py==2.6.4
ffmpeg-python==0.2.0
nextcord==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5